Disable edit & rejudge in proxy judge problems

+++ app/controllers/problems_controller.rb
@@ -33,6 +33,10 @@ class ProblemsController < ApplicationController
   end
 
   def rejudge
+    if @problem.proxyjudge_any?
+      redirect_back fallback_location: root_path, alert: 'This action is disabled in proxy judge problems.'
+      return
+    end
     subs = Submission.where(problem_id: params[:id], contest_id: @contest ? @contest.id : nil)
     sub_ids = subs.pluck(:id)
     SubmissionTestdataResult.where(submission_id: sub_ids).delete_all
